+++ 90s.sh
@@ -148,7 +148,7 @@ else
     src/90s/httpd/environment.cpp src/90s/httpd/render_context.cpp src/90s/httpd/server.cpp \
     src/90s/util/util.cpp \
     src/90s/sql/mysql.cpp src/90s/sql/mysql_util.cpp \
-    src/90s/mail/server.cpp src/90s/mail/indexed_mail_storage.cpp \
+    src/90s/mail/server.cpp src/90s/mail/parser.cpp src/90s/mail/indexed_mail_storage.cpp \
     src/90s/mail/http/http_api.cpp \
     src/90s/dns/doh.cpp \
     src/90s/mail/client.cpp
